Home

OMEinsum.jl

OMEinsum.Diag

OMEinsum.EinsumOp

OMEinsum.Fallback

OMEinsum.IndexReduction

OMEinsum.MixedDiag

OMEinsum.MixedStarContract

OMEinsum.OuterProduct

OMEinsum.Permutation

OMEinsum.StarContract

OMEinsum.TensorContract

OMEinsum.Trace

OMEinsum.bpcheck

OMEinsum.combineops

OMEinsum.edgesfrominds

OMEinsum.einsum

OMEinsum.einsum_grad

OMEinsum.einsumcost

OMEinsum.einsumopt

OMEinsum.evaluate

OMEinsum.evaluateall

OMEinsum.get_size_dict

OMEinsum.index_map

OMEinsum.indicesafteroperation

OMEinsum.iscombineable

OMEinsum.loop!

OMEinsum.opcost

OMEinsum.operatorfromedge

OMEinsum.operatorsfromedges

OMEinsum.optimalorder

OMEinsum.optimiseorder

OMEinsum.overlap

OMEinsum.einsum — Method.

einsum(ixs, xs, out)

return the tensor that results from contracting the tensors xs according to their indices ixs, where all indices that do not appear in the output are summed over. The indices are contracted in the order implied by their numerical value, smaller first. The result is permuted according to out.

ixs - tuple of tuple of integers that label all indices of a tensor. Indices that appear twice (in different tensors) are summed over

xs - tuple of tensors

out - tuple of integers that should correspond to remaining indices in ixs after contractions.

example

julia> a = rand(2,2);

julia> b = rand(2,2);

julia> einsum(((1,2),(2,3)), (a, b), (1,3)) ≈ a * b
true

julia> einsum(((1,2),(2,3)), (a, b), (3,1)) ≈ permutedims(a * b, (2,1))
true

source

OMEinsum.einsumopt — Method.

meinsumopt(ixs, xs, iy)

returns the result of the einsum operation implied by ixs, iy but evaluated in the optimal order according to meinsumcost.

source

OMEinsum.Diag — Type.

Diag{N,T}

is a type that represents a (generalized) diagonal of N edges of type T of one tensor which are stored in its edges field, e.g. 'iij -> ij' is represented by Diag{1}((i,))

source

OMEinsum.EinsumOp — Type.

EinsumOp{N}

abstract supertype of all einsum operations involving N edges or N tensors (for OuterProduct{N}).

source

OMEinsum.Fallback — Type.

Fallback{N,T}

is a type that represents an einsum resulting in N indices of type T, which are stored in its iy field. It's used as a general fallback if no more efficient method is available.

source

OMEinsum.IndexReduction — Type.

IndexReduction{N,T}

is a type that represents an index reduction of N edges/indices of type T which are stored in its edges field, e.g. 'ij -> i' is represented by IndexReduction{1,Char}((j,)).

source

OMEinsum.MixedDiag — Type.

MixedDiag{N,T}

is a type that represents a (generalized) mixed diagonal of N edges of type T of more than one tensor which are stored in its edges field, e.g. 'iij, ik -> ijk' is represented by MixedDiag{1,Char}((i,))

source

OMEinsum.MixedStarContract — Type.

MixedStarContract{N,T}

is a type that represents a mixed star-contraction of N edges of type T which are stored in its edges field. A MixedStarContract{N} results from N tensors sharing at least one index and at least one tensor has duplicate shared indices, e.g. 'ij,ik,iil -> jkl' is represented by MixedStarContract{1,Char}((i,)).

source

OMEinsum.OuterProduct — Type.

OuterProduct{N}

is a type that represents an outer product of N tensors.

source

OMEinsum.Permutation — Type.

Permutation{N,T}

is a type that represents a permutation of N indices which are stored in its perm field as a tuple of N integers of type T.

source

OMEinsum.StarContract — Type.

StarContract{N,T}

is a type that represents a star-contraction of N edges of type T which are stored in its edges field. A StarContract{N} results from N tensors sharing at least one index but no tensor has duplicate shared indices, e.g. 'ij,ik,il -> jkl' is represented by StarContract{1,Char}((i,)).

source

OMEinsum.TensorContract — Type.

TensorContract{N,T}

is a type that represents a tensorcontraction of N edges of type T which are stored in its edges field, e.g. 'ij,jk -> ik' is represented by TensorContract{1,Char}((j,)).

source

OMEinsum.Trace — Type.

Trace{N,T}

is a type that represents a trace operation of N edges of type T, i.e. 2N indices, which are stored in its edges field, e.g. 'ijjk -> ik' is represented by Trace{1,Char}((j,))

source

OMEinsum.bpcheck — Method.

bpcheck(f, args...; η = 1e-5, verbose=false)

returns a Bool indicating whether Zygote calculates the gradient of f(args...) -> scalar correctly using the relation f(x - ηg) ≈ f(x) - η|g|². If verbose=true, print f(x) - f(x - ηg)and η|g|².

source

OMEinsum.combineops — Method.

combineops(op1, op2)

return an operator that combines the operations of op1 and op2.

source

OMEinsum.edgesfrominds — Method.

edgesfrominds(ixs,iy)

return the edges of the ixs that imply an operation e.g. in ixs = ((1,2),(2,3)), iy = (1,3), edge 2 requires a tensor contraction

source

OMEinsum.einsum_grad — Method.

einsum_grad(ixs, xs, iy, y, i)

return gradient w.r.t the ith tensor in xs

source

OMEinsum.einsumcost — Method.

einsumcost(ixs, sxs, ops)

returns the cost of evaluating the einsum of ixs, sxs according to the sequence in ops, where sxs is a tuple of sizes.

source

OMEinsum.evaluate — Method.

evaluate(op::EinsumOp, allixs, allxs)

returns a tuple of xs and ixs that result from the evaluation of the operator op.

source

OMEinsum.evaluateall — Method.

evaluateall(ixs, xs, ops,iy)

evaluate the einsum specified by 'ixs -> iy' by going through all operations in ops in order.

source

OMEinsum.get_size_dict — Method.

get the dictionary of index=>size, error if there are conflicts

source

OMEinsum.index_map — Method.

take an index subset from ind

source

OMEinsum.indicesafteroperation — Method.

indicesafteroperation(op, allixs)

returns all indices of tensors after operation op was applied.

source

OMEinsum.iscombineable — Method.

iscombineable(a,b)

return true if EinsumOps a and b can be combined into one operator.

source

OMEinsum.loop! — Method.

loop and accumulate products to y

source

OMEinsum.opcost — Method.

opcost(op, ocost, allixs, allsxs)

returns the cost (in number of iterations it would require in a for loop) of evaluating op with arguments allixs and allsxs plus ocost as well as the new indices and sizes after evaluation.

allsxs is a tuple of tuples of Ints - the sizes of the respective arrays

source

OMEinsum.operatorfromedge — Method.

operatorfromedge(edge, ixs, iy)

returns a subtype of EinsumOp which specifies the kind of operation that the reduction of edge corresponds to.

source

OMEinsum.operatorsfromedges — Method.

operatorsfromedges(ixs, sxs, edges, iy)

given a list of edges edges, return a list of operations where consecutive operations are combined if possible.

source

OMEinsum.optimalorder — Method.

optimalorder(ixs, xs, iy)

return a tuple of operations that represents the (possibly nonunique) optimal order of reduction-operations.

source

OMEinsum.optimiseorder — Method.

optimiseorder(ixs, sxs, edges, iy)

return a tuple of operations that represents the (possibly nonunique) optimal order of reduction-operations in ops and its cost.

source

OMEinsum.overlap — Method.

overlap(s1, s2)

return true if s1 and s2 share any element.

source