Home

OMEinsum.jl

OMEinsum.contractcombine!

OMEinsum.diagonal

OMEinsum.diagonals

OMEinsum.einsum

OMEinsum.expandall!

OMEinsum.permuteandreshape

OMEinsum.einsum — Method.

einsum(cs, ts, out)

return the tensor that results from contracting the tensors ts according to their indices cs, where twice-appearing indices are contracted. The result is permuted according to out.

cs - tuple of tuple of integers that label all indices of a tensor. Indices that appear twice (in different tensors) are summed over

ts - tuple of tensors

out - tuple of integers that should correspond to remaining indices in cs after contractions.

This implementation has space requirements that are exponential in the number of unique indices.

example

julia> a = rand(2,2);

julia> b = rand(2,2);

julia> einsum(((1,2),(2,3)), (a, b), (1,3)) ≈ a * b
true

julia> einsum(((1,2),(2,3)), (a, b), (3,1)) ≈ permutedims(a * b, (2,1))
true

source

OMEinsum.expandall! — Method.

expandall!(b::AbstractArray{T,N}, indsb::NTuple{<:Any,N},
           a::AbstractArray{T,M}, indsa::NTuple{<:Any,M})

expands a into b by finding all index-labels in indsa in indsb and in case an label appears more in indsb than in indsa, the corresponding index is expanded. This enables operations such as

julia> OMEinsum.expandall!(zeros(2,2), (1,1), reshape([2]), ())
2×2 Array{Float64,2}:
 2.0  0.0
 0.0  2.0

where the zero-dimensional array reshape([2]) is expanded onto the diagonal of zeros(2,2) as in the backward action of taking the diagonal. Expansions are done by contracting a dirac-delta function with the indices to expand.

source

OMEinsum.contractcombine! — Method.

contractcombine!(out, outindspre, outinds, contractions, tensors)

take the tensorproduct of all tensors in tensors according to the specification in contractions and save the result in out.

source

OMEinsum.diagonal — Method.

diagonal(tensor::AbstractArray{<:Any,N}, inds::NTuple{N})

if there are any duplicate labels in inds, take those elements of tensor for which the indices for the duplicate labels are the same, i.e. if ind = (1,2,2), make a new tensor from the slices tensor[i,j,j]. If multiple duplicate labels are in inds recursively calls itself until none are left.

Example

julia> OMEinsum.diagonal([1 2; 3 4], (1,1))
([1, 4], (1,))

source

OMEinsum.diagonals — Method.

diagonals(tensors, contractions)

for each tensor in tensors, check whether there are any duplicates in contractions. If there are duplicates, take the diagonal w.r.t to the duplicates and return these diagonals and the new contractions.

Example

julia> OMEinsum.diagonals(([1 2; 3 4],), ((1,1),))
(([1, 4],), ((1,),))

source

OMEinsum.permuteandreshape — Method.

permuteandreshape(uniqueallins, tensor, inds)

permute tensor such that its indices in inds are in the same order as in uniqueallins. Then reshape the permuted tensor such that the indices of the resulting tensors for indices in inds is conserved while singleton-dimensions are inserted for indices in uniqueallins that are not in inds.

Example

Here, the array [1 2;3 4] is permuted according to the indices in uniqueallins, where label 1 comes before label 4 and then reshape it to have a shape that could be indexed with four labels.

julia> OMEinsum.permuteandreshape((1,2,3,4), [1 2; 3 4], (4,1))
2×1×1×2 Array{Int64,4}:
[:, :, 1, 1] =
 1
 2

[:, :, 1, 2] =
 3
 4

source