Home

OMEinsum.jl

OMEinsum.bpcheck

OMEinsum.einsum

OMEinsum.einsum_grad

OMEinsum.index_map

OMEinsum.loop!

OMEinsum.einsum — Method.

einsum(cs, ts, out)

return the tensor that results from contracting the tensors ts according to their indices cs, where twice-appearing indices are contracted. The result is permuted according to out.

cs - tuple of tuple of integers that label all indices of a tensor. Indices that appear twice (in different tensors) are summed over

ts - tuple of tensors

out - tuple of integers that should correspond to remaining indices in cs after contractions.

This implementation has space requirements that are exponential in the number of unique indices.

example

julia> a = rand(2,2);

julia> b = rand(2,2);

julia> einsum(((1,2),(2,3)), (a, b), (1,3)) ≈ a * b
true

julia> einsum(((1,2),(2,3)), (a, b), (3,1)) ≈ permutedims(a * b, (2,1))
true

source

OMEinsum.bpcheck — Method.

bpcheck(f, args...; η = 1e-5, verbose=false)

returns a Bool indicating whether Zygote calculates the gradient of f(args...) -> scalar correctly using the relation f(x - ηg) ≈ f(x) - η|g|². If verbose=true, print f(x) - f(x - ηg)and η|g|².

source

OMEinsum.einsum_grad — Method.

einsum_grad(ixs, xs, iy, y, i)

return gradient w.r.t the ith tensor in xs

source

OMEinsum.index_map — Method.

take an index subset from ind

source

OMEinsum.loop! — Method.

loop and accumulate products to y

source