OMEinsum.EinArray — Type

EinArray{T, N, TT, LX, LY, ICT, OCT} <: AbstractArray{T, N}

A struct to hold the intermediate result of an einsum where all index-labels of both input and output are expanded to a rank-N-array whose values are lazily calculated. Indices are arranged as inner indices (or reduced dimensions) first and then outer indices.

Type parameters are

* `T`: element type,
* `N`: array dimension,
* `TT`: type of "tuple of input arrays",
* `LX`: type of "tuple of input indexers",
* `LX`: type of output indexer,
* `ICT`: typeof inner CartesianIndices,
* `OCT`: typeof outer CartesianIndices,

OMEinsum.EinCode — Type

EinCode{ixs, iy}

Wrapper to eincode-specification that creates a callable object to evaluate the eincode ixs -> iy where ixs are the index-labels of the input-tensors and iy are the index-labels of the output

example

julia> a, b = rand(2,2), rand(2,2);

julia> EinCode((('i','j'),('j','k')),('i','k'))(a, b) ≈ a * b
true

OMEinsum.EinIndexer — Type

EinIndexer{locs,N}

A structure for indexing EinArrays. locs is the index positions (among all indices). In the constructor, size is the size of target tensor,

OMEinsum.EinIndexer — Method

EinIndexer{locs}(size::Tuple)

Constructor for EinIndexer for an object of size size where locs are the locations of relevant indices in a larger tuple.

OMEinsum.IndexGroup — Type

IndexGroup

Leaf in a contractiontree, contains the indices and the number of the tensor it describes, e.g. in "ij,jk -> ik", indices "ik" belong to tensor 1, so would be described by IndexGroup(['i','k'], 1).

OMEinsum.NestedEinsumConstructor — Type

NestedEinsumConstructor

describes a (potentially) nested einsum. Important fields:

args, vector of all inputs, either IndexGroup objects corresponding to tensors or NestedEinsumConstructor

iy, indices of output

Base.getindex — Method

getindex(A::EinArray, inds...)

return the lazily calculated entry of A at index inds.

OMEinsum.ContractionOrder.timespace_complexity — Method

ContractionOrder.timespace_complexity(eincode, size_dict)

Return the time and space complexity of the einsum contraction. The time complexity is defined as log2(number of element multiplication). The space complexity is defined as log2(size of the maximum intermediate tensor).

OMEinsum.allow_loops — Method

allow_loops(flag::Bool)

Setting this to false will cause OMEinsum to log an error if it falls back to loop_einsum evaluation, instead of calling specialised kernels. The default is true.

OMEinsum.allunique — Method

allunique(ix::Tuple)

return true if all elements of ix appear only once in ix.

example

julia> using OMEinsum: allunique

julia> allunique((1,2,3,4))
true

julia> allunique((1,2,3,1))
false

OMEinsum.analyze_binary — Method

Get the expected labels.

OMEinsum.asarray — Method

asarray(x[, parent::AbstractArray]) -> AbstactArray

Return a 0-dimensional array with item x, otherwise, do nothing. If a parent is supplied, it will try to match the parent array type.

OMEinsum.einarray — Method

einarray(::EinCode, xs, size_dict) -> EinArray

Constructor of EinArray from an EinCode, a tuple of tensors xs and a size_dict that assigns each index-label a size. The returned EinArray holds an intermediate result of the einsum specified by the EinCode with indices corresponding to all unique labels in the einsum. Reduction over the (lazily calculated) dimensions that correspond to labels not present in the output lead to the result of the einsum.

example

julia> using OMEinsum: get_size_dict

julia> a, b = rand(2,2), rand(2,2);

julia> sd = get_size_dict((('i','j'),('j','k')), (a, b));

julia> ea = OMEinsum.einarray(EinCode((('i','j'),('j','k')),('i','k')), (a,b), sd);

julia> dropdims(sum(ea, dims=1), dims=1) ≈ a * b
true

OMEinsum.einsum — Method

einsum(::EinCode{ixs, iy}, xs, size_dict) where {ixs, iy}

return the tensor that results from contracting the tensors xs according to their indices ixs, where all indices that do not appear in the output iy are summed over. The result is permuted according to out.

ixs - tuple of tuples of index-labels of the input-tensors xs

iy - tuple of index-labels of the output-tensor

xs - tuple of tensors

size_dict - a dictionary that maps index-labels to their sizes

example

julia> a, b = rand(2,2), rand(2,2);

julia> einsum(EinCode((('i','j'),('j','k')),('i','k')), (a, b)) ≈ a * b
true

julia> einsum(EinCode((('i','j'),('j','k')),('k','i')), (a, b)) ≈ permutedims(a * b, (2,1))
true

OMEinsum.einsum_grad — Method

einsum_grad(ixs, xs, iy, size_dict, cdy, i)

return the gradient of the result of evaluating the EinCode w.r.t the ith tensor in xs. cdy is the result of applying the EinCode to the xs.

example

julia> using OMEinsum: einsum_grad, get_size_dict

julia> a, b = rand(2,2), rand(2,2);

julia> c = einsum(EinCode((('i','j'),('j','k')), ('i','k')), (a,b));

julia> sd = get_size_dict((('i','j'),('j','k')), (a,b));

julia> einsum_grad((('i','j'),('j','k')), (a,b), ('i','k'), sd, c, 1) ≈ c * transpose(b)
true

OMEinsum.filliys! — Method

filliys!(neinsum::NestedEinsumConstructor)

goes through all NestedEinsumConstructor objects in the tree and saves the correct iy in them.

OMEinsum.get_size_dict — Function

get_size_dict(ixs, xs, size_info=nothing)

return a dictionary that is used to get the size of an index-label in the einsum-specification with input-indices ixs and tensors xs after consistency within ixs and between ixs and xs has been verified.

OMEinsum.indices_and_locs — Method

indices_and_locs(ixs,iy)

given the index-labels of input and output of an einsum, return (in the same order):

a tuple of the distinct index-labels of the output iy

a tuple of the distinct index-labels in ixs of the input not appearing in the output iy

a tuple of tuples of locations of an index-label in the ixs in a list of all index-labels

a tuple of locations of index-labels in iy in a list of all index-labels

where the list of all index-labels is simply the first and the second output catenated and the second output catenated.

OMEinsum.loop_einsum! — Method

loop_einsum!(::EinCode, xs, y, size_dict)

inplace-version of loop_einsum, saving the result in a preallocated tensor of correct size y.

OMEinsum.loop_einsum — Method

loop_einsum(::EinCode, xs, size_dict)

evaluates the eincode specified by EinCode and the tensors xs by looping over all possible indices and calculating the contributions ot the result. Scales exponentially in the number of distinct index-labels.

OMEinsum.map_prod — Method

map_prod(xs, ind, indexers)

calculate the value of an EinArray with EinIndexers indexers at location ind.

OMEinsum.match_rule — Method

match_rule(ixs, iy)
match_rule(code::EinCode{ixs, iy})
match_rule(code::NestedEinCode)

Returns the rule that matches, otherwise use DefaultRule - the slow loop_einsum backend.

OMEinsum.nopermute — Method

nopermute(ix,iy)

check that all values in iy that are also in ix have the same relative order,

example

julia> using OMEinsum: nopermute

julia> nopermute((1,2,3),(1,2))
true

julia> nopermute((1,2,3),(2,1))
false

e.g. nopermute((1,2,3),(1,2)) is true while nopermute((1,2,3),(2,1)) is false

OMEinsum.optimize_greedy — Method

optimize_greedy(eincode, size_dict; method=MinSpaceOut(), nrepeat=10)

Greedy optimizing the contraction order and return a NestedEinsum object. Methods are

MinSpaceOut, always choose the next contraction that produces the minimum output tensor.

MinSpaceDiff, always choose the next contraction that minimizes the total space.

OMEinsum.parse_parens — Method

parse_parens(s::AbstractString, i, narg)

parse one level of parens starting at index i where narg counts which tensor the current group of indices, e.g. "ijk", belongs to. Recursively calls itself for each new opening paren that's opened.

OMEinsum.tensorpermute — Method

tensorpermute(A, perm)

Aliasing permutedims(A, perm).

OMEinsum.@ein — Macro

@ein A[i,k] := B[i,j] * C[j,k]     # A = B * C

Macro interface similar to that of other packages.

You may use numbers in place of letters for dummy indices, as in @tensor, and need not name the output array. Thus A = @ein [1,2] := B[1,ξ] * C[ξ,2] is equivalent to the above. This can also be written A = ein"ij,jk -> ik"(B,C) using the numpy-style string macro.

example

julia> a, b = rand(2,2), rand(2,2);

julia> @ein c[i,k] := a[i,j] * b[j,k];

julia> c ≈ a * b
true

julia> c ≈ ein"ij,jk -> ik"(a,b)
true

OMEinsum.@ein_str — Macro

ein"ij,jk -> ik"(A,B)

String macro interface which understands numpy.einsum's notation. Translates strings into EinCode-structs that can be called to evaluate an einsum. To control evaluation order, use parentheses - instead of an EinCode, a NestedEinsum is returned which evaluates the expression according to parens. The valid character ranges for index-labels are a-z and α-ω.

example

julia> a, b, c = rand(10,10), rand(10,10), rand(10,1);

julia> ein"ij,jk,kl -> il"(a,b,c) ≈ ein"(ij,jk),kl -> il"(a,b,c) ≈ a * b * c
true